+++ NEWS.md
@@ -1,3 +1,5 @@
+# hubAdmin (development version)
+
 # hubAdmin 1.8.0
 
 * Added basic tests for v6.0.0 schema (#127).